requirements.txt
dm-haiku==0.0.10
optax==0.1.7
distrax==0.1.3
tensorflow-probability==0.22.0

gymnasium==0.29.0
gymnasium[box2d]==0.29.0
gymnasium[mujoco]==0.29.0
robosuite==1.4.0
dm_control==1.0.14

numpy==1.22.4
pandas==2.0.3
tensorboard==2.13.0
tensorboardx==2.6.2
cpprb==10.7.1
h5py==3.9.0
git+https://github.com/Farama-Foundation/Metaworld.git@04be337a12305e393c0caf0cbf5ec7755c7c8feb
recordtype==1.4
rliable==1.0.8
termcolor==2.3.0
cython==0.29.36
